backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
django-cors-headers>=4.0.0
Pillow>=9.0.0
